client/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
contourpy==1.2.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
fonttools==4.56.0
    # via matplotlib
grpcio==1.71.0
    # via grpcio-tools
grpcio-tools==1.71.0
    # via -r requirements.in
kiwisolver==1.4.8
    # via matplotlib
matplotlib==3.8.4
    # via -r requirements.in
numpy==1.26.4
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
packaging==24.2
    # via matplotlib
pillow==11.1.0
    # via matplotlib
polars==1.26.0
    # via -r requirements.in
protobuf==5.29.4
    # via grpcio-tools
pyparsing==3.2.1
    # via matplotlib
python-dateutil==2.9.0.post0
    # via matplotlib
six==1.17.0
    # via python-dateutil

# The following packages are considered to be unsafe in a requirements file:
# setuptools
